// src/types/bookmark.ts
export interface BookMarkProps {
    post: {
        id: number;
        genre: string;
        postUrl: string;
        ageRating: string;
        title: string;
        startDate: string;
        endDate: string;
        operatingHours: string;
        location: string;
        venue: string;
        status: string;
        ticketingWebSite: string;
        price: string;
        detailImage: string;
        description: string;
        bookmarked: boolean;
    };
    
  }
  